requirements.txt
python-telegram-bot
requests
pony
apscheduler
qrcode
Pillow
httpx